+++ example/source/main.cpp
@@ -1,5 +1,5 @@
 #define TESLA_INIT_IMPL // If you have more than one file using the tesla header, only define this in the main one
-#include <tesla_extended_components.hpp>    // The Tesla Header for extended components, use <tesla.hpp> is you need only standard components
+#include <tesla.hpp>    // The Tesla Header
 
 
 class GuiSecondary : public tsl::Gui {
